Object_Detection_YOLOv8/requirements.txt
opencv-python
opencv-python-headless
Pillow
tk
ultralytics
cvzone
numpy
openai
